Add rate limits for restore
